+++ CHANGELOG.md
@@ -1,3 +1,44 @@
+<a name="2.0.195"></a>
+## [2.0.195](https://github.com/bullhorn/novo-elements/compare/v2.0.194...v2.0.195) (2017-05-16)
+
+
+### Bug Fixes
+
+* **aot:** Fixing AOT build for FileInput ([c05a149](https://github.com/bullhorn/novo-elements/commit/c05a149))
+
+
+
+<a name="2.0.194"></a>
+## [2.0.194](https://github.com/bullhorn/novo-elements/compare/v2.0.193...v2.0.194) (2017-05-16)
+
+
+### Features
+
+* **checkbox:** make icon optional - style can be box or circle ([#476](https://github.com/bullhorn/novo-elements/issues/476)) ([f6dd554](https://github.com/bullhorn/novo-elements/commit/f6dd554))
+* **Table:** Added Expand all details feature ([#477](https://github.com/bullhorn/novo-elements/issues/477)) ([6a933a8](https://github.com/bullhorn/novo-elements/commit/6a933a8))
+
+
+
+<a name="2.0.193"></a>
+## [2.0.193](https://github.com/bullhorn/novo-elements/compare/v2.0.192...v2.0.193) (2017-05-12)
+
+
+### Bug Fixes
+
+* **appbridge:** Fixing HTTP calls via app bridge ([26cabd5](https://github.com/bullhorn/novo-elements/commit/26cabd5))
+
+
+
+<a name="2.0.192"></a>
+## [2.0.192](https://github.com/bullhorn/novo-elements/compare/v2.0.191...v2.0.192) (2017-05-12)
+
+
+### Bug Fixes
+
+* **table:** Fix instance where selectAll was not showing ([e7a6b7c](https://github.com/bullhorn/novo-elements/commit/e7a6b7c))
+
+
+
 <a name="2.0.191"></a>
 ## [2.0.191](https://github.com/bullhorn/novo-elements/compare/v2.0.190...v2.0.191) (2017-05-10)